+++ src/components/AddLink.jsx
@@ -1,0 +1,5 @@
+const AddLink = () => {
+  return <></>;
+};
+
+export default AddLink;
