+++ src/components/FolderInfo.jsx
@@ -1,5 +1,20 @@
-const FolderInfo = ({ data, name }) => {
-  console.log(data);
+import { useEffect, useState } from 'react';
+import { getFolderData } from '../apis/api';
+
+const FolderInfo = () => {
+  const [data, setData] = useState({});
+  const [name, setName] = useState('');
+
+  const fetchFolderData = async () => {
+    const res = await getFolderData();
+    setName(res.folder.name);
+    setData(res.folder.owner);
+  };
+
+  useEffect(() => {
+    fetchFolderData();
+  }, []);
+
   return (
     <div>
       <div>{data.name}</div>
