+++ src/components/SearchBar.jsx
@@ -1,0 +1,5 @@
+const SearchBar = () => {
+  return <></>;
+};
+
+export default SearchBar;
