+++ src/pages/folder.jsx
@@ -1,0 +1,16 @@
+import AddLink from '../components/AddLink';
+import NavBar from '../components/NavBar/NavBar';
+import SearchBar from '../components/SearchBar';
+import '../styles/global.css';
+
+const Folder = () => {
+  return (
+    <>
+      <NavBar />
+      <AddLink />
+      <SearchBar />
+    </>
+  );
+};
+
+export default Folder;
